Move `index.js` to `index.web.js`
