+++ packages/webpack-config/tests/nextjs/pages/index.js
@@ -1,23 +1,6 @@
-import { LinearGradient } from 'expo-linear-gradient';
 import React from 'react';
-import { StyleSheet, Text } from 'react-native';
-import Constants from 'expo-constants';
-import getenv from 'getenv';
+import { Text } from 'react-native';
 
 export default function App() {
-  return (
-    <LinearGradient colors={['orange', 'blue']} style={styles.container}>
-      <Text testID="basic-text">Open up App.js to start working on your app!</Text>
-      <Text testID="expo-constants-manifest">{JSON.stringify(Constants.manifest)}</Text>
-      {getenv.boolish('CI', false) && <Text testID="has-ci-text">Has CI env</Text>}
-    </LinearGradient>
-  );
+  return <Text>This is for the native app.</Text>;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
